+++ src/components/ui/button/index.tsx
@@ -1,5 +1,5 @@
+import { cn } from '@libs/shadcn';
 import { Slot } from '@radix-ui/react-slot';
-import { cn } from '@utils/shadcn';
 import { cva, type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
