+++ src/components/ui/shadcn-ui/checkbox.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import { cn } from '@libs/shadcn';
 import * as CheckboxPrimitive from '@radix-ui/react-checkbox';
-import { cn } from '@utils/shadcn';
 import { Check } from 'lucide-react';
 import * as React from 'react';
