+++ src/components/ui/textarea/index.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@utils/shadcn';
+import { cn } from '@libs/shadcn';
 import * as React from 'react';
 
 export interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {}
