+++ src/components/ui/toast/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import { cn } from '@libs/shadcn';
 import * as ToastPrimitives from '@radix-ui/react-toast';
-import { cn } from '@utils/shadcn';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { X } from 'lucide-react';
 import * as React from 'react';
